requirements.txt
flask==2.0.1
google-cloud-storage==2.10.0
google-cloud-bigquery==3.11.4
pre-commit
black
